Removing the trips screen now
